@@ -22,15 +22,13 @@ android {
         abortOnError false
     }
     compileOptions {
-        targetCompatibility JavaVersion.VERSION_1_7
-        sourceCompatibility JavaVersion.VERSION_1_7
+        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_1_8
     }
 }
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support:recyclerview-v7:27.+'
-    implementation 'com.android.support:cardview-v7:27.+'
     //noinspection GradleCompatible
     implementation 'com.google.firebase:firebase-config:11.6.2'
     implementation 'com.google.firebase:firebase-auth:11.6.2'
@@ -45,11 +43,10 @@ dependencies {
     //noinspection GradleCompatible
     implementation 'com.android.support:recyclerview-v7:27.1.1'
     implementation 'com.android.support:design:27.1.1'
+    implementation 'com.android.support:cardview-v7:27.1.1'
     implementation 'com.squareup.okhttp3:okhttp:3.10.0'
-    implementation 'com.google.code.gson:gson:2.8.0'
+
     implementation 'com.ibm.watson.developer_cloud:conversation:4.2.1'
-    implementation 'com.ibm.watson.developer_cloud:text-to-speech:4.2.1'
-    implementation 'com.ibm.watson.developer_cloud:speech-to-text:4.2.1'
     implementation 'com.ibm.watson.developer_cloud:android-sdk:0.4.3'
     implementation 'com.ibm.watson.developer_cloud:java-sdk:6.6.0'
     implementation 'com.android.support:multidex:1.0.1'
