@@ -5,15 +5,16 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
 public class CheckingHomework {
-    private static DatabaseReference myRef = FirebaseDatabase.getInstance().getReference();
+    private String userUid = FirebaseAuth.getInstance().getUid();
+    private  DatabaseReference myRef = FirebaseDatabase.getInstance().getReference("users/" + userUid);
 
     // 체크 박스 누를 시에 과제 내의 isChecked 를 변화 시킴
-    public static void setChecking(String homeworkUid, Boolean integer){
-        if(FirebaseAuth.getInstance().getCurrentUser()!=null){
+    public void setChecking(String homeworkUid, Boolean bool) {
+        if (FirebaseAuth.getInstance().getCurrentUser() != null) {
             String userUid = FirebaseAuth.getInstance().getUid();
-            if(userUid != null && !userUid.equals(""))
-                myRef.child("homeWorks").child(userUid).child(homeworkUid).child("isChecked").setValue(integer);
-        }else{
+            if (userUid != null && !userUid.equals(""))
+                myRef.child("tasks").child(homeworkUid).child("checked").setValue(bool);
+        } else {
             System.out.println("error");
         }
     }
