@@ -40,7 +40,7 @@ public class HomeworkAdapter extends RecyclerView.Adapter<HomeworkAdapter.Review
     @Override
     public HomeworkAdapter.ReviewViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         // 뷰 홀더 생성시 호출 어떻게 생성해서 무엇을 전달할 것인가. 새롭게 생성할 때만 호출된다.
-        View itemView = LayoutInflater.from(parent.getContext()).inflate(R.layout.homework_card_show, parent, false);
+        View itemView = LayoutInflater.from(parent.getContext()).inflate(R.layout.homework_item, parent, false);
         return new HomeworkAdapter.ReviewViewHolder(itemView);
     }
 
@@ -52,8 +52,14 @@ public class HomeworkAdapter extends RecyclerView.Adapter<HomeworkAdapter.Review
     }
 
     public static class ReviewViewHolder extends RecyclerView.ViewHolder {
-        @BindView(R.id.homework_goal)
-        TextView homework_goal;
+        @BindView(R.id.homework_cognitive_error)
+        TextView hwCognitiveError;
+        @BindView(R.id.homework_date)
+        TextView hwDate;
+        @BindView(R.id.homework_automatic_thought)
+        TextView hwAutoThought;
+        @BindView(R.id.homework_assignment)
+        TextView hwAssignment;
         @BindView(R.id.homework_finished_button)
         CheckBox homework_finished_button;
         @BindView(R.id.homework_form)
@@ -71,10 +77,13 @@ public class HomeworkAdapter extends RecyclerView.Adapter<HomeworkAdapter.Review
 
         private void setData(Homework newHomework, String uid) {
             // Insert Data
-            homework_goal.setText(newHomework.getGoal());
+            hwCognitiveError.setText(newHomework.getCognitiveError());
+            hwDate.setText(newHomework.getWrittenTime());
+            hwAutoThought.setText(newHomework.getAutomaticThought());
+            hwAssignment.setText(newHomework.getAssignment());
             homework_uid.setText(uid);
             // layout to check box button
-            if (newHomework.getIsChecked()) {
+            if (newHomework.getChecked()) {
                 homework_finished_button.setChecked(true);
                 visRemoveButton();
             } else {
@@ -86,19 +95,20 @@ public class HomeworkAdapter extends RecyclerView.Adapter<HomeworkAdapter.Review
 
         @OnCheckedChanged(R.id.homework_finished_button)
         public void isChecking(boolean isChecked) {
+            CheckingHomework checkingHomework = new CheckingHomework();
             if (isChecked) {
                 // 체크 하면
                 Log.w("finish", "완료");
                 // 불투명 하게
                 visRemoveButton();
                 // 체크했다고 디비에 연락
-                CheckingHomework.setChecking(homework_uid.getText().toString(), true);
+                checkingHomework.setChecking(homework_uid.getText().toString(), true);
             } else {
                 // 체크 해제하면
                 Log.w("release", "해제");
                 noRemoveButton();
                 // 체크 헤제 했다고 디비에 연락
-                CheckingHomework.setChecking(homework_uid.getText().toString(), false);
+                checkingHomework.setChecking(homework_uid.getText().toString(), false);
             }
         }
 
