@@ -7,22 +7,21 @@ import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.Query;
 import com.google.firebase.database.ValueEventListener;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class HomeworkReader {
-    private static DatabaseReference myRef = FirebaseDatabase.getInstance().getReference();
-    private static String userUid = FirebaseAuth.getInstance().getUid();
+    private String userUid = FirebaseAuth.getInstance().getUid();
+    private DatabaseReference myRef = FirebaseDatabase.getInstance().getReference("users/" + userUid);
 
     private static List<Homework> homeworkList;
     private static List<String> homeworkUidList;
 
-    public static void getInitData(final HomeworkListener homeworkListener) {
+    public void getInitData(final HomeworkListener homeworkListener) {
 
-        myRef.child("homeWorks").child(userUid).addValueEventListener(new ValueEventListener() {
+        myRef.child("tasks").addValueEventListener(new ValueEventListener() {
             @Override
             public void onDataChange(DataSnapshot dataSnapshot) {
 
@@ -35,14 +34,17 @@ public class HomeworkReader {
                     // init
                     Homework newHomework = new Homework();
                     // add data
-                    if (getHomework != null)
-                        newHomework = new Homework(getHomework.getGoal(), getHomework.getStartTime(), getHomework.getEndTime(), getHomework.getIsChecked());
-
+                    if (getHomework != null) {
+                        newHomework = new Homework(getHomework.getAssignment()
+                                , getHomework.getCognitiveError()
+                                , getHomework.getAutomaticThought()
+                                , getHomework.getChecked()
+                                , getHomework.getWrittenTime());
+                    }
                     // add to ArrayList
                     homeworkList.add(newHomework);
                     homeworkUidList.add(data.getKey());
                 }
-
                 homeworkListener.goalListener(homeworkList, homeworkUidList);
             }
 
